+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased Changes
 - Update the `model_training_xgboost.py` documentation.
+- Add the `CrossValidationMetricsOperator` to compute scoring for all partitions for a model.
+- Add the `ScoreBacktestsModelOperator` to score all backtests for a datetime partitioned model.
 
 
 ## 0.3.0